Step 1: remove the pure-Ruby encoder, and use the JSON gem instead
Ideally, we want something like this:

module ActiveSupport
  module JSON
    def self.encode(value, options = nil)
      JSON.generate(value.as_json(options), quirks_mode: true)
    end
  end
end

Unfortunately, it doesn't get THAT simple, because we need to support
some exising features, namely:

1. Some non-standard string encoding (AS wants \u2028 and \u2029
   escaped)

2. The "escape" HTML entities feature

3. Encoding BigDecimal as a JSON numeric value (currently via an ugly
   hack)

So I kept the stuff required to made this work, all existing tests
are passing. Also took advantage of the UTF-8 stuff in Ruby 1.9+ to
simplify the escaping stuff.