Added some failing tests for as_json not being called recursively
As @jeremy pointed out in #11728 and on campfire, as_json is meant to
be a hook for "hinting" the JSON encoder on how to encode the object
into JSON, it's the encoder's responsibility to recurrsively call
as_json on the returned value. These tests points out what aren't
working currently.

(Note: some of these currently fail on master as well)